resolve six easy warnings from clang-tidy

+++ core/loader/TiffDataReader.cpp
@@ -86,7 +86,7 @@ Eigen::MatrixXi TiffDataReader::data(std::size_t /*frame*/)
         Eigen::Matrix<uint16, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> data16(
             _nRows, _nCols);
         // Read line per line
-        for (unsigned short int i = 0; i < _nRows; ++i)
+        for (size_t i = 0; i < _nRows; ++i)
             TIFFReadScanline(_file, (char*)&data16(i, 0), i);
         // Not very nice, but need to copy the 16bits data to int
         return data16.cast<int>();
@@ -94,7 +94,7 @@ Eigen::MatrixXi TiffDataReader::data(std::size_t /*frame*/)
         Eigen::Matrix<uint32, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> data32(
             _nRows, _nCols);
         // Read line per line
-        for (unsigned short int i = 0; i < _nRows; ++i)
+        for (size_t i = 0; i < _nRows; ++i)
             TIFFReadScanline(_file, (char*)&data32(i, 0), i);
         // Not very nice, but need to copy the 32bits data to int
         return data32.cast<int>();
